+++ shaders/vert-default.glsl
@@ -1,7 +1,7 @@
-#version 450 core
+#version 460 core
 layout (location = 0) in vec4 aPos;
 layout (location = 1) in vec4 aCol;
-layout (location = 3) uniform vec4 viewport;
+layout (location = 3) uniform uvec4 viewport;
 out vec4 vsColor;
 
 void main()
